[BEAM-5605] Increase precision of fraction used during splitting.

+++ sdks/go/pkg/beam/core/runtime/exec/plan.go
@@ -201,7 +201,7 @@ func (p *Plan) Metrics() *fnpb.Metrics {
 type SplitPoints struct {
 	// Splits is a list of desired split indices.
 	Splits []int64
-	Frac   float32
+	Frac   float64
 }
 
 // Split takes a set of potential split indexes, and if successful returns
